refac : 리폼의뢰 페이지 폼 수정
리폼 의뢰 페이지 폼을 상품 등록,수정 폼과 동일하게 변경

#Issues 74
