+++ src/component/TopBar.js
@@ -18,11 +18,7 @@ export default function TopBar() {
     if (accessToken) {
       setIsLoggedIn(true);
       axiosInstance
-        .get("/user/role", {
-          headers: {
-            accessToken
-          },
-        })
+        .get("/user/role")
         .then(async (response) => {
           setRole(response.data.data.role);
           localStorage.setItem("role", response.data.data.role);
@@ -37,6 +33,7 @@ export default function TopBar() {
 
   const handleLogout = () => {
     localStorage.removeItem("accessToken");
+    document.cookie = 'accessToken=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;';
     clearTokens()
     setIsLoggedIn(false);
     navigate("/login");
