+++ src/pages/main.js
@@ -6,6 +6,7 @@ import { useNavigate,Link } from 'react-router-dom';
 import TopBar from "../component/TopBar.js";
 import { useEndPointStore } from '../store.js';
 import { TailWindButton } from '../component/atoms/Button.js';
+import { useTokenStore } from '../store.js';
 
 
 function Main(){
@@ -13,6 +14,7 @@ function Main(){
     const [sort, setSort] = useState(true);
     const [productDesc,setProductDesc] = useState();
     const {Endpoint} = useEndPointStore(state => state)
+    const accessToken = useTokenStore((state) => state.accessToken)
     
     useEffect(() => {
         const fetchData = async () => {
